+++ R/restructure.R
@@ -166,6 +166,10 @@ restructure_v2 <- function(x) {
   if (version < "2.12.11") {
     # added support for 'cmdstanr' as additional backend
     x$backend <- "rstan"
+    # argument 'position' was added to stanvars
+    for (i in seq_along(x$stanvars)) {
+      x$stanvars[[i]]$position <- "start"
+    }
   }
   x
 }